scraper-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
playwright==1.49.1
python-dotenv==1.0.1
requests==2.31.0
google-auth==2.28.0
